moved update methods from sensor to TadoDataStore
