+++ homeassistant/components/tado_v1.py
@@ -9,11 +9,13 @@ import logging
 import urllib
 
 import voluptuous as vol
+from datetime import timedelta
 
-from homeassistant.components.discovery import load_platform
+from homeassistant.helpers.discovery import load_platform
 from homeassistant.helpers import config_validation as cv
 from homeassistant.const import (
     CONF_USERNAME, CONF_PASSWORD)
+from homeassistant.util import Throttle
 
 
 _LOGGER = logging.getLogger(__name__)
@@ -35,6 +37,8 @@ CONFIG_SCHEMA = vol.Schema({
     })
 }, extra=vol.ALLOW_EXTRA)
 
+MIN_TIME_BETWEEN_SCANS = timedelta(seconds=10)
+
 
 def setup(hass, config):
     """Your controller/hub specific code."""
@@ -49,19 +53,61 @@ def setup(hass, config):
         _LOGGER.error("Unable to connect to mytado with username and password")
         return False
 
-    hass.data['tado_v1_data'] = TadoDataStore(tado)
+    hass.data['tado_v1_data'] = TadoDataStore(tado, MIN_TIME_BETWEEN_SCANS)
 
     for component in TADO_V1_COMPONENTS:
         load_platform(hass, component, DOMAIN, {}, config)
 
     return True
 
+
 class TadoDataStore:
     """An object to store the tado data."""
 
-    def __init__(self, tado):
+    def __init__(self, tado, interval):
         """Initialize Tado data store."""
         self.tado = tado
+
+        self.sensors = {}
+        self.data = {}
+
+        # Apply throttling to methods using configured interval
+        self.update = Throttle(interval)(self._update)
+
+    def _update(self):
+        """Update the internal data from mytado.com."""
+        for data_id, sensor in self.sensors.items():
+            data = None
+
+            try:
+                if "zone" in sensor:
+                    _LOGGER.info("querying mytado.com for zone %s %s",
+                                 sensor["id"], sensor["name"])
+                    data = self.tado.getState(sensor["id"])
+                if "device" in sensor:
+                    _LOGGER.info("querying mytado.com for device %s %s",
+                                 sensor["id"], sensor["name"])
+                    data = self.tado.getDevices()[0]
+
+            except RuntimeError:
+                _LOGGER.error("Unable to connect to myTado. %s %s",
+                              sensor["id"], sensor["id"])
+
+            self.data[data_id] = data
+
+    def add_sensor(self, data_id, sensor):
+        """Add a sensor to update in _update()."""
+        self.sensors[data_id] = sensor
+        self.data[data_id] = None
+
+    def get_data(self, data_id):
+        """Get the cached data."""
+        data = {"error": "no data"}
+
+        if data_id in self.data:
+            data = self.data[data_id]
+
+        return data
 
     def get_zones(self):
         """Wrapper for getZones()."""
@@ -78,7 +124,7 @@ class TadoDataStore:
     def reset_zone_overlay(self, zone_id):
         """Wrapper for resetZoneOverlay(..)."""
         return self.tado.resetZoneOverlay(zone_id)
-    
+
     def set_zone_overlay(self, zone_id, mode, temperature):
         """Wrapper for setZoneOverlay(..)."""
         return self.tado.setZoneOverlay(zone_id, mode, temperature)
